Split copy mode motion commands into separate file
